pyproject.toml
[project]
name = "fitness-coach-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "python-telegram-bot[all]==20.7",
    "schedule>=1.2.0",
    "sqlalchemy>=2.0.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.0",
    "oauthlib>=3.2.2",
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "trafilatura>=2.0.0",
    "pandas>=2.0.0",
    "twilio>=9.0.0"
]

[tool.poetry.dependencies]
yookassa = "^2.4.0"
